Crate forest_interpreter

Structs

BlockMessages

Contains all messages to process through the VM as well as miner information for block rewards.

FixedRewardCalc

Giving a fixed amount of coins for each block produced directly to the miner, on top of the gas spent, so the circulating supply isn’t burned. Ignores penalties.

NoRewardCalc

Not giving any reward for block creation.

RewardActorMessageCalc

Default reward working with the Filecoin Reward Actor.

VM

Interpreter which handles execution of state transitioning messages and returns receipts from the VM execution.

Traits

RewardCalc

Allows the generation of a reward message based on gas fees and penalties.

Functions

resolve_to_key_addr

returns the public key type of address (BLS/SECP256K1) of an account actor identified by addr.