Trait forest_interpreter::RewardCalc

pub trait RewardCalc: Send + Sync + 'static {
    fn reward_message(
        &self,
        epoch: ChainEpoch,
        miner: Address,
        win_count: i64,
        penalty: TokenAmount,
        gas_reward: TokenAmount
    ) -> Result<Option<Message>, Error>;
}

Allows the generation of a reward message based on gas fees and penalties.

This should facilitate custom consensus protocols using their own economic incentives.

Required Methods

fn reward_message(
&self,
epoch: ChainEpoch,
miner: Address,
win_count: i64,
penalty: TokenAmount,
gas_reward: TokenAmount
) -> Result<Option<Message>, Error>

Construct a reward message, if rewards are applicable.